requirements.txt
lark==1.3.0
lxml==6.0.2
motor==3.7.1
orjson==3.11.3
pyjwt[crypto]==2.10.1
pymongo==4.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
sanic[ext]==25.3.0
tqdm==4.67.1
